Tighten PairStream and Combination matching queries
Refine the JPQL used to match PairStream and Combination rows so that
set equality is expressed more explicitly and deterministically.

Without this change, concurrent creates could legitimately produce
multiple rows representing the same logical PairStream or Combination,
causing lookup queries to return more than one result and fail with
non-unique result errors, or select an arbitrary matching row.

This change does not introduce uniqueness guarantees; it simply makes
best-effort reuse stable and predictable when duplicates already exist.

+++ backend/src/main/kotlin/dev/coldhands/pair/stairs/backend/infrastructure/persistance/dao/JpaCombinationEventDao.kt
@@ -92,7 +92,7 @@ class JpaCombinationEventDao(
             pairStreamRepository.findByDevelopersAndStream(
                 pairStream.developerIds.map { it.value }.toSet(),
                 pairStream.streamId.value,
-                pairStream.developerIds.size
+                pairStream.developerIds.size.toLong()
             )
                 .firstOrNull()
                 ?: pairStreamRepository.save(
@@ -109,7 +109,7 @@ class JpaCombinationEventDao(
 
         val combinationEntity: CombinationEntity = combinationRepository.findByPairStreams(
             pairStreamEntities.map { it.id!! },
-            pairStreamEntities.size
+            pairStreamEntities.size.toLong()
         )
             .firstOrNull()
             ?: combinationRepository.save(
